+++ packages/analytics_hub/lib/src/core/interception/dispatch/event_dispatcher.dart
@@ -48,7 +48,7 @@ class EventDispatcher {
         properties: event.properties,
         context: event.context,
       ),
-      target.eventProvider.options,
+      target.eventProvider.overrides,
     );
     final interceptors = [
       ..._hubInterceptors,