+++ packages/analytics_hub_firebase/lib/src/provider/firebase_analytics_hub_provider.dart
@@ -18,7 +18,7 @@ class FirebaseAnalyticsHubProvider extends AnalytycsProvider {
   FirebaseAnalyticsHubProvider({required FirebaseAnalytics analytics})
       : _analytics = analytics,
         super(
-          identifier: FirebaseAnalyticsHubProviderIdentifier(
+          identifier: FirebaseAnalyticsHubIdentifier(
             name: analytics.app.name,
           ),
           interceptors: const [],