+++ packages/analytics_hub_firebase/test/src/resolver/firebase_analytics_event_resolver_test.dart
@@ -70,7 +70,7 @@ class _TestEvent extends Event {
   @override
   List<EventProvider> get providers => [
         const EventProvider(
-          FirebaseAnalyticsHubProviderIdentifier(name: 'test'),
+          FirebaseAnalyticsHubIdentifier(name: 'test'),
         ),
       ];
 }